refactor(discord-orchestrator): split into apps/delulu_discord + apps/delulu_sandbox_modal
Separate the VPS-hosted Discord bot and the Modal sandbox function into
two fully independent sub-projects under apps/. Each owns its own
pyproject.toml, uv.lock, .python-version, Makefile, and (for the bot)
Dockerfile. The Discord bot's modal_dispatch/sandbox.py is renamed to
delulu_discord/dispatcher.py to make explicit that it's a client-side
wrapper around modal.Function.from_name, not a thing that runs inside
the sandbox. The top-level Makefile becomes a thin dispatcher over the
two sub-project Makefiles.

Verified: both sub-projects uv sync, import, and pass ruff check +
format --check locally; Modal App name still resolves to
'discord-orchestrator' so Function.from_name lookups in the bot stay
unchanged.
