+++ infra/discord-orchestrator/apps/delulu_discord/.python-version
