+++ infra/discord-orchestrator/apps/delulu_discord/src/delulu_discord/handlers.py
@@ -8,9 +8,9 @@ import discord
 import structlog
 
 if TYPE_CHECKING:
-    from src.bot.session_manager import SessionManager
-    from src.config import Settings
-    from src.modal_dispatch.sandbox import SandboxDispatcher
+    from delulu_discord.dispatcher import SandboxDispatcher
+    from delulu_discord.session_manager import SessionManager
+    from delulu_discord.settings import Settings
 
 logger = structlog.get_logger()
