+++ infra/discord-orchestrator/apps/delulu_sandbox_modal/.python-version
@@ -1,0 +1,1 @@
+3.14
