+++ infra/discord-orchestrator/apps/delulu_sandbox_modal/pyproject.toml
@@ -1,0 +1,29 @@
+[project]
+name = "delulu-sandbox-modal"
+version = "0.1.0"
+description = "Modal sandbox function that runs Claude Code for the delulu Discord bot"
+requires-python = ">=3.14"
+dependencies = [
+    "modal>=1.0",
+    "structlog>=24.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "ruff>=0.15",
+    "pytest>=8.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/delulu_sandbox_modal"]
+
+[tool.ruff]
+target-version = "py314"
+line-length = 100
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "N", "W", "UP"]
