--- infra/discord-orchestrator/src/config/__init__.py
@@ -1,3 +1,0 @@
-from src.config.settings import Settings
-
-__all__ = ["Settings"]
